Set karpenter provisioner `ttlSecondsAfterEmpty`.
Our prow infra utilizes karpenter to scale up when there is the need of
execution multiple prow jobs in parallel (making new releases for
examples). However we observed that once those jobs are finished, the
karpenter provisioned nodes are note quickly drained and stay attached
approximatively an hour.. which increases the cost of our infra
unecessarily.

This patch sets a `ttlSecondsAfterEmpty` to instruct karpenter to drain
and delete nodes if they are not running any workload for 15minutes.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>