Set karpenter provisioner `ttlSecondsAfterEmpty` to 15minutes (#439)
Our prow infra utilizes karpenter to scale up when there is the need of
execution multiple prow jobs in parallel (making new releases for
examples). However we observed that once those jobs are finished, the
karpenter provisioned nodes are note quickly drained and stay attached
approximatively an hour.. which increases the cost of our infra
unecessarily.

This patch sets a `ttlSecondsAfterEmpty` to instruct karpenter to drain
and delete nodes if they are not running any workload for 15minutes.

This patch also removes the tiny/memory optimized ec2 machine types
since most of our prow job pods require 8CPU/3Gi of memory.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.