Install libffi for ACM in pytest image (#459)
Description of changes:
As part of https://github.com/aws-controllers-k8s/acm-controller/pull/40, an integration test was added that has a dependency on the `cryptography` module. In order for this to work, `libffi` needs to be installed in Alpine-based images.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.