Add upgrade-go-version prow  job and patch prow image versions (#465)
Description of changes:
This PR adds a new periodic Prow job `upgrade-go-version` that checks the latest Go version available in the ECR public gallery and compares it with the version used in our repository (`build_config.yaml`). If a newer Go version is available, it:

1. Updates the `build_config.yaml` file with the latest Go version.
2. Bumps the patch versions of all Prow job images in `images_config.yaml`.
3. Commits the changes to a new branch.
4. Raises this PR against the `image-config-update` branch.

This automation helps us keep our Go version up-to-date and ensures that Prow job images are rebuilt whenever there is a new Go version, without requiring manual intervention. 

The next prow job that will be pushed will read the PR pushed by upgrade-go-version and automatically build and push prow images to ECR. 

In the future we plan to include more build versions that we want to periodically query, and upgrade if needed, to ensure our prow jobs dependencies are always up-to-date, without requiring any human intervention.

TODO for this PR:
- [x] Add upgrade-go-version to images_config.yaml
- [x] Create Dockerfile.upgrade-go-version
- [x] Add command to periodic jinja 
- [ ] Use go_version in build_config.yaml in existing Dockerfiles

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.