Add a new prow job to automatically build `test-infra` images (inception) (#466)
Description of changes:

This PR introduces a new Prow postsubmit job `build-prow-images` that runs after a commit makes a change to `images_config.yaml` in the `test-infra` repository. The job is responsible for keeping the Prow job images up-to-date by rebuilding them if a the version in `images_config.yaml` is greater than the one in ECR public. If the process is successful, it makes a PR with a re-generated `jobs.yaml` file

The purpose of this PR addresses two issues:
1. This would automate building and pushing prow job images to ECR, eliminating human intervention, making sure our prow images are always up-to-date
2. It uses kaniko to build and push images in the cluster, since Dockershim is will no longer be supported in our cluster after September 1st.

TODO:

- [x]  Add patch-prow-jobs to images_config.yaml
- [x] Add binary build in Makefile
- [x] Create Dockerfile.patch-prow-jobs
- [x] Add command to postsubmit jinja
- [ ] Regenerate jobs.yaml and create a PR

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.