updating olm-pr-bundle to add ci.yaml file, if it is found missing (#503)
Issue #, if available:
NA

Description of changes:
Change to create the controller directory and `ci.yaml` file, if it is found missing, instead of creating a GitHub issue. The tools used are already in the `image` that this job uses:

https://github.com/aws-controllers-k8s/test-infra/blob/1dba4bd0ba9008814297646d83310c02365dca2a/prow/jobs/images/Dockerfile.olm-bundle-pr#L70-L72

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.