Add `upgrade-eks-distro-version` prow job
This prow job acts similarly to `upgrade-go-version`.
With this prow job, we will ensure that we have the latest
build images, and in the future we will add any other dependencies
we want to keep an eye on.

This PR also contains some refactoring done to `upgrade-go-version`
to allow reusability.