Add `upgrade-eks-distro-version` prow job (#524)
Description of changes:

This prow job acts similarly to `upgrade-go-version`. With this prow job, 
we will ensure that we have the latest build images, and in the future we 
will add any other dependencies we want to keep an eye on.

This PR also contains some refactoring done to `upgrade-go-version` to allow reusability.
We are also using the `upgrade-go-version` image for this job.

In the future we will see if we can combine these two jobs into one

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.