Bump prow integrations image version
Looks like we have already built v0.0.23 in the past, trying to upgrade
the kind version, but we failed miserably and had to rollback to v0.0.22.

Recently `ack-build-tools` bumped all the versions, but it didn't force push
a new image for our integrations tests (as intended). Now we want to trigger
a new build for the ingrations image.