Bump prow integrations image version (#529)
It looks like we have already built `v0.0.23` in the past, trying to upgrade
the kind version, but we failed miserably and had to rollback to `v0.0.22`.

Recently `ack-build-tools` bumped all the versions, but it didn't force push
a new image for our integrations tests (as intended). Now we want to trigger
a new build for the ingrations image.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.